# perf: skip tracestate regex validation for es vendor key